refac(consts): separated terrain from weather
also added weather labels for use in FieldCalc weather dropdown options
