+++ src/consts/field/index.ts
@@ -1,1 +1,2 @@
+export * from './terrain';
 export * from './weather';
